Round-robin scheduler
The round-robin scheduler is a continuation that “wraps” several other
child continuations.  When the scheduler is executed, it executes all of
its child continuations in the order that they were added to the
scheduler.  When all of the children have completed (by transferring
control to `cps_done`), the scheduler itself completes, passing control
to whichever `next` continuation was provided when the scheduler was
started.  The children are allowed to add additional entries to the
scheduler while it's executing.