cps_rr_run_one_lap function
The new `cps_rr_run_one_lap` function executes all of the continuations
that are currently in the round-robin scheduler's queue, and then
returns.  If any additional continuations are added during that
execution, they'll be remembered, but won't be executed until the next
call.

The `cps_rr_run` function has been renamed to `cps_rr_drain`; it runs
*all* of the continuations in the queue, even those that are added
during the execution.