Work around CMake 2.6's assembly “support”
CMake 2.6's assembly support is still experimental.  It ends up using
too many of the C compiler's flags when constructing the assembler's
command line.  This patch overrides the assembler command-line template
to only pass in the options that the assembler will recognize.  (But we
only do this when using CMake 2.6, since CMake 2.8 and higher have nicer
support for assembler-specific options.)