Fix include directories for context detector
We have some cmake magic that detects which context implementation to
use, using libcork's platform detection macros to do the hard work.
That means we need to be able to find libcork's header files; if they're
not installed into a standard system location, then we need this new
clause to find them.